requirements.txt
annotated-types==0.5.0
appdirs==1.4.4
awscli==1.29.84
clip==1.0
cryptography==41.0.3
detectron2==0.6
diffusers==0.21.2
easydict==1.9
fqdn==1.5.1
frontier-exploration==0.1
frozendict==2.3.8
glide-text2im==0.0.0
gputil==1.4.0
groundingdino==0.1.0
home-robot==0.1.0
huggingface-hub>=0.17.0
img2dataset==1.42.0
imgaug==0.4.0
isoduration==20.11.0
jsonpointer==2.4
jupyter==1.0.0
kornia==0.5.0
lazy-loader==0.3
linkify-it-py==2.0.2
llava==1.1.1
magnum==0.0.0
mss==9.0.1
natsort==8.4.0
open-clip-torch>=2.20.0
patchify==0.2.3
prior==1.0.3
procthor==0.0.1.dev2
pydantic-core==2.6.3
pynput==1.7.6
pytest==7.4.2
pytorch-lightning==1.2.9
pytorch-triton==2.1.0+6e4932cda8
salesforce-lavis==1.0.2
scikit-fmm==2023.4.2
seaborn==0.12.2
segment-anything==1.0
tensorboard-plugin-wit==1.8.1
tensorboardx==2.6.2.2
tensorflow==2.13.0
tiktoken==0.5.2
transformers>=4.36.0
ttach==0.0.3
uri-template==1.3.0
wavedrom==2.0.3.post3
webcolors==1.13
wldhx.yadisk-direct==0.0.6
